requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.1.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
tqdm>=4.66.0

# Data processing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Configuration
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Logging
colorlog>=6.7.0

# Development
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0